always align segmentation of circles
Rounded end segments were converted to a polygon using
a specified "aCircleToSegmentCount". But the segmentation
was not aligned to be 0 degree, delta degree, 2*delta degree etc,
but it was aligned with the angle of the segment.
Often, two rounded segment ends are joined (boolAdd).
Because their segmentation did not align, this resulted
in more segments than necessary and sometimes very small segments.
This commit changes this behavior to create aligned
segmentations.